+++ core/src/main/kotlin/fragments/AndroidNavigationNodeFactory.kt
@@ -14,16 +14,18 @@ class AndroidNavigationNodeFactory<T : NavigationNodeDefaultConfig>(
     private val flowOnHierarchyChangeListener: FlowOnHierarchyChangeListener = FlowOnHierarchyChangeListener(recursive = true).also {
         (rootView as ViewGroup).setOnHierarchyChangeListenerRecursively(it)
     },
+    private val manualHierarchyCheckerDelayMillis: Long? = 100L,
     private val fragmentKClassResolver: FragmentsClassesFactory<T>
 ) : NavigationNodeFactory<T> {
-    override fun createNode(navigationChain: NavigationChain<T>, config: T): NavigationNode<T>? {
+    override fun createNode(navigationChain: NavigationChain<T>, config: T): NavigationNode<out T, T>? {
         return AndroidFragmentNode(
             navigationChain,
             config,
             fragmentKClassResolver(config) ?: return null,
             fragmentManager,
             rootView,
-            flowOnHierarchyChangeListener
+            flowOnHierarchyChangeListener,
+            manualHierarchyCheckerDelayMillis
         )
     }
 }