+++ mvvm/src/commonMain/kotlin/View.kt
@@ -1,6 +1,6 @@
 package dev.inmo.navigation.mvvm
 
-interface View {
-    fun draw()
-    fun clear()
-}
+//interface View {
+//    fun draw()
+//    fun clear()
+//}