+++ mvvm/src/commonMain/kotlin/ViewModel.kt
@@ -2,4 +2,4 @@ package dev.inmo.navigation.mvvm
 
 import dev.inmo.navigation.core.NavigationNode
 
-abstract class ViewModel<T> : NavigationNode<T>()
+//abstract class ViewModel<T> : NavigationNode<T>()